requirements.txt
torch>=2.0.0
numpy
pandas
gymnasium
minigrid
stable-baselines3
transformers
accelerate
bitsandbytes
peft
trl
tensorboard
pyyaml
protobuf==4.25.8
sentence-transformers
